requirements.txt
django
djangorestframework
requests
djlint